Fix : HubController로 합침
