+++ hub-service/src/main/java/com/gbg/hubservice/application/service/HubService.java
@@ -3,6 +3,7 @@ package com.gbg.hubservice.application.service;
 import com.gbg.hubservice.domain.entity.Hub;
 import com.gbg.hubservice.presentation.dto.request.CreateHubRequestDto;
 import com.gbg.hubservice.presentation.dto.request.UpdateHubRequestDto;
+import com.gbg.hubservice.presentation.dto.response.GetHubResponseDto;
 import java.util.UUID;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -18,4 +19,6 @@ public interface HubService {
     void update(UUID hubId, UpdateHubRequestDto request);
 
     void delete(UUID hubId, UUID actorId);
+
+    GetHubResponseDto getByUserId(UUID hubManagerId);
 }
